Extract cso dr management into new spec
